backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
qdrant-client
langchain
langchain-community
langchain-huggingface
langchain-qdrant
sentence-transformers